move framework to proper location, general cleanup
