+++ src/main/framework/settings.ts
