+++ src/main/framework/workspace.ts
@@ -2,12 +2,12 @@ import { join, resolve } from 'path'
 import { Settings } from './settings'
 import { mkdirs, pathExists } from 'fs-extra'
 import { homedir } from 'node:os'
-import { MTermWindow } from '../main/window/mterm-window'
+import { MTermWindow } from '../window/mterm-window'
 import { remove } from 'lodash'
 import { BrowserWindowConstructorOptions } from 'electron'
-import { setWindowValueFromPath } from '../main/util'
+import { setWindowValueFromPath } from '../util'
 import { Runtime } from './runtime'
-import { DEFAULT_FOLDER } from '../constants'
+import { DEFAULT_FOLDER } from '../../constants'
 import { Commands } from './commands'
 
 export function resolveFolderPathForMTERM(folder: string): string {
