+++ tsconfig.node.json
@@ -1,6 +1,7 @@
 {
   "extends": "@electron-toolkit/tsconfig/tsconfig.node.json",
-  "include": ["electron.vite.config.*", "src/main/**/*", "src/preload/**/*", "src/framework/**/*", "src/constants.ts"],
+  "include": ["electron.vite.config.*", "src/main/**/*", "src/preload/**/*",
+    "src/main/framework/**/*", "src/constants.ts"],
   "exclude": [
     "src/**/*.test.ts"
   ],
